pyproject.toml
[project]
name = "core-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi>=0.136.1",
    "pydantic-settings>=2.14.0",
    "pyjwt[crypto]>=2.12.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn>=0.46.0",
]
